requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
faiss-cpu
pypdf
python-dotenv
google-generativeai
python-docx
python-pptx
pandas
sentence-transformers